Note to connect CLI to an instance before deploy
This could perhaps be re-worded, but the gist is that my co-workers and I were running into issues with the `ent bundle deploy` command because we hadn't seen anything about attaching the Entando VM while following the other setup tutorials. Running `ent attach-vm entando` at the suggestion of @nshaw fixed the problem for us, so I thought this note might save future developers a bit of searching.